ssz: change class names so pre-sos implementation match old names
