+++ ssz/src/main/java/org/ethereum/beacon/ssz/visitor/SSZSimpleSerializer.java
@@ -11,119 +11,94 @@ import java.io.ByteArrayOutputStream;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.ethereum.beacon.ssz.visitor.SSZSimpleDeserializer.BYTES_PER_LENGTH_OFFSET;
+public class SSZSimpleSerializer implements SSZVisitor<SSZSimpleSerializer.SSZHasherSerializerResult, Object>  {
 
-public class SSZSimpleSerializer
-    implements SSZVisitor<SSZSimpleSerializer.SSZSerializerResult, Object> {
-  private static BytesValue serializeLength(long len) {
-    return BytesValues.ofUnsignedIntLittleEndian(len);
-  }
+  public static class SSZHasherSerializerResult {
+    BytesValue serializedBody;
+    BytesValue serializedLength;
 
-  @Override
-  public SSZSerializerResult visitBasicValue(SSZBasicType type, Object value) {
-    ByteArrayOutputStream baos = new ByteArrayOutputStream();
-    type.getAccessor().encode(value, type.getTypeDescriptor(), baos);
-    return new SSZSerializerResult(
-        BytesValue.wrap(baos.toByteArray()), baos.size(), type.isFixedSize());
-  }
-
-  @Override
-  public SSZSerializerResult visitList(
-      SSZListType type, Object param, ChildVisitor<Object, SSZSerializerResult> childVisitor) {
-
-    if (type.isVector()) {
-      if (type.getChildrenCount(param) != type.getVectorLength()) {
-        throw new SSZSerializeException(
-            "Vector type length doesn't match actual list length: "
-                + type.getVectorLength()
-                + " !=  "
-                + type.getChildrenCount(param)
-                + " for "
-                + type.toStringHelper());
-      }
-    }
-    return visitComposite(type, param, childVisitor);
-  }
-
-  @Override
-  public SSZSerializerResult visitSubList(
-      SSZListType type,
-      Object param,
-      int startIdx,
-      int len,
-      ChildVisitor<Object, SSZSerializerResult> childVisitor) {
-    return visitComposite(type, param, childVisitor, startIdx, len);
-  }
-
-  @Override
-  public SSZSerializerResult visitComposite(
-      SSZCompositeType type,
-      Object rawValue,
-      ChildVisitor<Object, SSZSerializerResult> childVisitor) {
-    return visitComposite(type, rawValue, childVisitor, 0, type.getChildrenCount(rawValue));
-  }
-
-  private SSZSerializerResult visitComposite(
-      SSZCompositeType type,
-      Object rawValue,
-      ChildVisitor<Object, SSZSerializerResult> childVisitor,
-      int startIdx,
-      int len) {
-
-    List<SSZSerializerResult> childSerializations = new ArrayList<>();
-    boolean fixedSize = type.isFixedSize();
-    for (int i = startIdx; i < startIdx + len; i++) {
-      SSZSerializerResult res = childVisitor.apply(i, type.getChild(rawValue, i));
-      childSerializations.add(res);
+    public SSZHasherSerializerResult(BytesValue serializedBody) {
+      this.serializedBody = serializedBody;
+      this.serializedLength = BytesValue.EMPTY;
     }
 
-    int currentOffset =
-        childSerializations.stream()
-            .mapToInt(r -> r.isFixedSize() ? r.serializedBody.size() : BYTES_PER_LENGTH_OFFSET)
-            .sum();
-    BytesValue composite = BytesValue.EMPTY;
-
-    // Fixed part
-    for (SSZSerializerResult s : childSerializations) {
-      composite =
-          composite.concat(s.isFixedSize() ? s.serializedBody : serializeLength(currentOffset));
-      if (!s.isFixedSize()) {
-        currentOffset = currentOffset + s.serializedLength;
-      }
-    }
-
-    // Variable part
-    for (SSZSerializerResult s : childSerializations) {
-      if (s.isFixedSize()) {
-        continue;
-      }
-      composite = composite.concat(s.serializedBody);
-    }
-
-    return new SSZSerializerResult(composite, currentOffset, fixedSize);
-  }
-
-  public static class SSZSerializerResult {
-    final BytesValue serializedBody;
-    final int serializedLength;
-    final boolean fixedSize;
-
-    public SSZSerializerResult(BytesValue serializedBody, int serializedLength, boolean fixedSize) {
+    public SSZHasherSerializerResult(BytesValue serializedBody,
+                               BytesValue serializedLength) {
       this.serializedBody = serializedBody;
       this.serializedLength = serializedLength;
-      this.fixedSize = fixedSize;
     }
 
     public BytesValue getSerializedBody() {
       return serializedBody;
     }
 
-    public int getSerializedLength() {
+    public BytesValue getSerializedLength() {
       return serializedLength;
     }
 
-    public boolean isFixedSize() {
-      return fixedSize;
+    public BytesValue getSerialized() {
+      return BytesValue.concat(getSerializedLength(), getSerializedBody());
     }
+
+    public boolean isFixedSize() {
+      return serializedLength.isEmpty();
+    }
+  }
+
+  @Override
+  public SSZHasherSerializerResult visitBasicValue(SSZBasicType type, Object value) {
+    ByteArrayOutputStream baos = new ByteArrayOutputStream();
+    type.getAccessor().encode(value, type.getTypeDescriptor(), baos);
+    BytesValue res = BytesValue.wrap(baos.toByteArray());
+    if (!type.isFixedSize()) {
+      res = serializeLength(res.size()).concat(res);
+    }
+    return new SSZHasherSerializerResult(res);
+  }
+
+  @Override
+  public SSZHasherSerializerResult visitList(SSZListType type, Object param,
+                                                           ChildVisitor<Object, SSZHasherSerializerResult> childVisitor) {
+
+    if (type.isVector()) {
+      if (type.getChildrenCount(param) != type.getVectorLength()) {
+        throw new SSZSerializeException("Vector type length doesn't match actual list length: "
+            + type.getVectorLength() + " !=  " + type.getChildrenCount(param) + " for " + type.toStringHelper());
+      }
+    }
+    return visitComposite(type, param, childVisitor);
+  }
+
+  @Override
+  public SSZHasherSerializerResult visitSubList(SSZListType type, Object param,
+                                                              int startIdx, int len, ChildVisitor<Object, SSZHasherSerializerResult> childVisitor) {
+    return visitComposite(type, param, childVisitor, startIdx, len);
+  }
+
+  @Override
+  public SSZHasherSerializerResult visitComposite(SSZCompositeType type, Object rawValue,
+                                                                ChildVisitor<Object, SSZHasherSerializerResult> childVisitor) {
+    return visitComposite(type, rawValue, childVisitor, 0, type.getChildrenCount(rawValue));
+  }
+
+  private SSZHasherSerializerResult visitComposite(SSZCompositeType type, Object rawValue,
+                                                                 ChildVisitor<Object, SSZHasherSerializerResult> childVisitor, int startIdx, int len) {
+    List<BytesValue> childSerializations = new ArrayList<>();
+    boolean fixedSize = type.isFixedSize();
+    int length = 0;
+    for (int i = startIdx; i < startIdx + len; i++) {
+      SSZHasherSerializerResult res = childVisitor.apply(i, type.getChild(rawValue, i));
+      childSerializations.add(res.serializedLength);
+      childSerializations.add(res.serializedBody);
+      fixedSize &= res.isFixedSize();
+      length += res.serializedBody.size() + res.serializedLength.size();
+    }
+
+    return new SSZHasherSerializerResult(BytesValue.concat(childSerializations),
+        fixedSize ? BytesValue.EMPTY : serializeLength(length));
+  }
+
+  private BytesValue serializeLength(int len) {
+    return BytesValues.ofUnsignedIntLittleEndian(len);
   }
 }
